+++ ChangeLog.md
@@ -11,7 +11,9 @@
 
 ### Deprecated methods
 
-- .
+- [[MASDevice currentDevice] logOutDeviceAndClearLocal:completion:] is deprecated to avoid confusion from developers and to align with Android SDK.  Please use [[MASUser currentUser] logoutWithCompletion:] method to log-out the authenticated user.
+- [[L7SClientManager sharedManager] logoutDevice] is also completely deprecated from the backward-compatibility for above reason. 
+- [[MASUser currentUser] logoffWithCompletion:] was renamed to [[MASUser currentUser] logoutWithCompletion:].
 
 
 # Version 1.1.00